[refactor] 전체적인 클래스 이름, 메서드 이름 리팩터링
