--- src/main/java/io/github/petty/pipeline/service/PromptAssemblyService.java
@@ -1,5 +1,0 @@
-package io.github.petty.pipeline.service;
-
-public interface PromptAssemblyService {
-    String assemblePrompt(String visionReport, String location) throws Exception;
-}
