+++ src/main/java/io/github/petty/pipeline/service/TogetherServiceImpl.java
@@ -18,13 +18,17 @@ public class TogetherServiceImpl implements TogetherService {
     private String apiUrl;
     @Value("${together.api.key}")
     private String apiKey;
+    @Value("${together.api.model}")
+    private String apiModel;
 
     @Override
     public String answer(String prompt) throws Exception {
+
         HttpClient client = HttpClient.newHttpClient();
         ObjectMapper mapper = new ObjectMapper();
-        String model = "meta-llama/Llama-3.3-70B-Instruct-Turbo-Free";
-        String body = mapper.writeValueAsString(new TogetherRequestDTO(model, List.of(new TogetherRequestDTO.Message("user", prompt))));
+
+        String body = mapper.writeValueAsString(new TogetherRequestDTO(apiModel, List.of(new TogetherRequestDTO.Message("user", prompt))));
+
         HttpRequest request = HttpRequest.newBuilder()
                 .uri(URI.create("%s".formatted(apiUrl)))
                 .header("Authorization", "Bearer %s".formatted(apiKey))
@@ -36,6 +40,6 @@ public class TogetherServiceImpl implements TogetherService {
         if (response.statusCode() == 200) {
             return mapper.readValue(response.body(), TogetherResponseDTO.class).choices().get(0).message().content();
         }
-        throw new Exception("잘못된 결과.");
+        throw new Exception("Together API를 호출하지 못하였습니다.");
     }
 }
