+++ src/main/java/io/github/petty/pipeline/util/TogetherPromptBuilder.java
@@ -1,0 +1,5 @@
+package io.github.petty.pipeline.util;
+
+public interface TogetherPromptBuilder {
+    String buildPrompt(String visionReport, String location) throws Exception;
+}
